feat(net): add net module

+++ kernel/src/fs/devfs/cpu_dma_latency.rs
@@ -79,12 +79,12 @@ impl File for LatencyFile {
             let _sum_guard = SumGuard::new();
             buf.fill(0);
             debug!("/dev/cpu_dma_latency: fill 0");
-            Ok(buf.len() as isize)
+            Ok(buf.len())
         })
     }
     fn write<'a>(&'a self, buf: &'a [u8]) -> AsyscallRet {
         debug!("[write] /dev/cpu_dma_latency");
-        Box::pin(async move { Ok(buf.len() as isize) })
+        Box::pin(async move { Ok(buf.len()) })
     }
 
     fn sync_read(&self, buf: &mut [u8]) -> SyscallRet {
@@ -92,12 +92,12 @@ impl File for LatencyFile {
         let _sum_guard = SumGuard::new();
         buf.fill(0);
         debug!("[sync_read] /dev/cpu_dma_latency: fill 0");
-        Ok(buf.len() as isize)
+        Ok(buf.len())
     }
 
     fn sync_write(&self, buf: &[u8]) -> SyscallRet {
         debug!("[sync_write] /dev/cpu_dma_latency");
-        Ok(buf.len() as isize)
+        Ok(buf.len())
     }
     fn flags(&self) -> OpenFlags {
         self.meta.inner.lock().flags
