+++ kernel/src/fs/devfs/rtc.rs
@@ -79,12 +79,12 @@ impl File for RtcFile {
             let _sum_guard = SumGuard::new();
             buf.fill(0);
             debug!("/dev/rtc: fill 0");
-            Ok(buf.len() as isize)
+            Ok(buf.len())
         })
     }
     fn write<'a>(&'a self, buf: &'a [u8]) -> AsyscallRet {
         debug!("write /dev/rtc");
-        Box::pin(async move { Ok(buf.len() as isize) })
+        Box::pin(async move { Ok(buf.len()) })
     }
     fn flags(&self) -> OpenFlags {
         self.meta.inner.lock().flags
