+++ kernel/src/fs/procfs/mounts.rs
@@ -79,7 +79,7 @@ impl File for MountsFile {
             } else {
                 buf[..len].copy_from_slice(info.as_bytes());
                 inner.pos = len;
-                Ok(len as isize)
+                Ok(len)
             }
         })
     }
